Change option's position
