+++ package.json
@@ -15,7 +15,6 @@
     "@types/three": "^0.126.2",
     "axios": "^0.21.0",
     "firebase": "^8.3.1",
-    "html-to-image": "^1.3.24",
     "html2canvas": "^1.0.0-rc.7",
     "immer": "^8.0.0",
     "lodash": "^4.17.20",
@@ -25,7 +24,6 @@
     "react": "^16.14.0",
     "react-dom": "^16.14.0",
     "react-hotkeys-hook": "^3.3.0",
-    "react-spinners": "^0.10.6",
     "react-spring": "^9.0.0",
     "react-three-fiber": "^5.3.22",
     "react-use-measure": "^2.0.4",
